New naming scheme for components
