+++ example-app/app.mandates.js
@@ -1,6 +1,6 @@
 import path from 'path';
 
-const entryModulePath = path.resolve(__dirname, 'index.module.js');
+const entryModulePath = path.resolve(__dirname, 'index-module.js');
 const entryTemplatePath = path.resolve(__dirname, 'index.html');
 const entryModuleName = 'IndexModule';
 const appDirectory = path.resolve(__dirname, 'app');
