+++ example-app/app/app-component.js
@@ -1,0 +1,5 @@
+import PartyMember from '../lib/party-member';
+
+const AppComponent = new PartyMember('app-component', {}, {}, 'app/app-component.html');
+
+export default AppComponent;
