+++ example-app/app/date/date-component.html
