--- example-app/app/date/date.component.js
@@ -1,5 +1,0 @@
-import PartyMember from '../../lib/party-member';
-
-const DateComponent = new PartyMember('date-component', { date: '2021-02-01' }, {}, 'app/date/date.component.html');
-
-export default DateComponent;
