+++ example-app/app/select/select-component.html
