+++ example-app/index-module.js
@@ -1,0 +1,13 @@
+import AppComponent from './app/app-component';
+import DateComponent from './app/date/date-component';
+import MainComponent from './app/main/main-component';
+import SelectComponent from './app/select/selectcomponent';
+
+const IndexModule = [
+    AppComponent,
+    DateComponent,
+    MainComponent,
+    SelectComponent
+];
+
+export { IndexModule };
